+++ polymorphia-frontend/components/grading-components/grade/criteria/types.ts
@@ -1,0 +1,5 @@
+import { CriterionResponseDTO } from "@/interfaces/api/grade/criteria";
+
+export interface GradeCriteriaProps {
+  criteria: CriterionResponseDTO[];
+}
