--- polymorphia-frontend/components/grading-components/grade/index.css
@@ -1,26 +1,0 @@
-@import "tailwindcss";
-@import "../../../styles/globals.css";
-
-.grade-accordion-override {
-    @apply gap-y-5;
-}
-
-.grade-criterion {
-    @apply flex flex-col mt-2;
-
-    h2 {
-        @apply text-3xl my-3;
-    }
-}
-
-.grade-criterion-progress-bar {
-    @apply w-full px-6 pt-6;
-}
-
-.grade-input-wrapper {
-    @apply w-20 mx-auto flex items-center -mt-6 z-[20];
-}
-
-.grade-comment-wrapper {
-    @apply w-full mt-2;
-}
